+++ mmselfsup/core/optimizer/__init__.py
@@ -2,5 +2,9 @@
 from .builder import build_optimizer
 from .constructor import DefaultOptimizerConstructor
 from .optimizers import LARS
+from .transformer_finetune_constructor import TransformerFinetuneConstructor
 
-__all__ = ['LARS', 'build_optimizer', 'DefaultOptimizerConstructor']
+__all__ = [
+    'LARS', 'build_optimizer', 'TransformerFinetuneConstructor',
+    'DefaultOptimizerConstructor'
+]