+++ mmselfsup/datasets/data_sources/__init__.py
@@ -3,5 +3,10 @@ from .base import BaseDataSource
 from .cifar import CIFAR10, CIFAR100
 from .image_list import ImageList
 from .imagenet import ImageNet
+from .imagenet_21k import ImageNet21k
+from .breastscreening import NYUBreastScreening
 
-__all__ = ['BaseDataSource', 'CIFAR10', 'CIFAR100', 'ImageList', 'ImageNet']
+__all__ = [
+    'BaseDataSource', 'CIFAR10', 'CIFAR100', 'ImageList', 'ImageNet',
+    'ImageNet21k', 'NYUBreastScreening'
+]