+++ mmselfsup/datasets/pipelines/__init__.py
@@ -1,5 +1,8 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .transforms import (GaussianBlur, Lighting, RandomAppliedTrans,
-                         Solarization)
 
-__all__ = ['GaussianBlur', 'Lighting', 'RandomAppliedTrans', 'Solarization']
+from .transforms import (BlockwiseMaskGenerator, GaussianBlur, Lighting, 
+                         RandomAppliedTrans, RandomAug, Solarization, 
+                         ToNumpy, Standardizer, CopyChannel)
+
+__all__ = ['GaussianBlur', 'Lighting', 'RandomAppliedTrans', 'Solarization',
+    'RandomAug', 'BlockwiseMaskGenerator', 'ToNumpy', 'Standardizer', 'CopyChannel']