+++ mmselfsup/models/algorithms/__init__.py
@@ -2,8 +2,11 @@
 from .base import BaseModel
 from .byol import BYOL
 from .classification import Classification
+from .breast_classification import USClassification
 from .deepcluster import DeepCluster
 from .densecl import DenseCL
+from .mae import MAE
+from .mmcls_classifier_wrapper import MMClsImageClassifierWrapper
 from .moco import MoCo
 from .mocov3 import MoCoV3
 from .npid import NPID
@@ -11,11 +14,14 @@ from .odc import ODC
 from .relative_loc import RelativeLoc
 from .rotation_pred import RotationPred
 from .simclr import SimCLR
+from .simmim import SimMIM
 from .simsiam import SimSiam
 from .swav import SwAV
+from .breast_swav import BreastSwAV
 
 __all__ = [
     'BaseModel', 'BYOL', 'Classification', 'DeepCluster', 'DenseCL', 'MoCo',
-    'MoCoV3', 'NPID', 'ODC', 'RelativeLoc', 'RotationPred', 'SimCLR',
-    'SimSiam', 'SwAV'
+    'NPID', 'ODC', 'RelativeLoc', 'RotationPred', 'SimCLR', 'SimSiam', 'SwAV',
+    'MAE', 'MoCoV3', 'SimMIM', 'MMClsImageClassifierWrapper', 'BreastSwAV',
+    'USClassification'
 ]