+++ tests/test_models/test_algorithms/test_classification.py
@@ -1,9 +1,13 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import platform
+
+import pytest
 import torch
 
 from mmselfsup.models.algorithms import Classification
 
 
+@pytest.mark.skipif(platform.system() == 'Windows', reason='Windows mem limit')
 def test_classification():
     # test ResNet
     with_sobel = True,
@@ -23,8 +27,10 @@ def test_classification():
 
     fake_input = torch.randn((16, 3, 224, 224))
     fake_labels = torch.ones(16, dtype=torch.long)
-    fake_backbone_out = alg.extract_feat(fake_input)
-    assert fake_backbone_out[0].size() == torch.Size([16, 2048, 7, 7])
+    fake_out = alg.forward_test(fake_input)
+    assert 'head4' in fake_out
+    assert fake_out['head4'].size() == torch.Size([16, 4])
+
     fake_out = alg.forward_train(fake_input, fake_labels)
     assert fake_out['loss'].item() > 0
 
@@ -43,7 +49,7 @@ def test_classification():
     )
 
     alg = Classification(backbone=backbone, head=head)
-    assert hasattr(alg, 'head')
+    assert alg.with_head
 
     fake_input = torch.randn((16, 3, 224, 224))
     fake_labels = torch.ones(16, dtype=torch.long)