+++ tests/test_models/test_algorithms/test_deepcluster.py
@@ -1,4 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import platform
+
 import pytest
 import torch
 
@@ -20,6 +22,7 @@ head = dict(
     num_classes=num_classes)
 
 
+@pytest.mark.skipif(platform.system() == 'Windows', reason='Windows mem limit')
 def test_deepcluster():
     with pytest.raises(AssertionError):
         alg = DeepCluster(
@@ -33,7 +36,10 @@ def test_deepcluster():
 
     fake_input = torch.randn((16, 3, 224, 224))
     fake_labels = torch.ones(16, dtype=torch.long)
-    fake_backbone_out = alg.extract_feat(fake_input)
-    assert fake_backbone_out[0].size() == torch.Size([16, 2048, 7, 7])
+    fake_out = alg.forward(fake_input, mode='test')
+    assert 'head0' in fake_out
+    assert fake_out['head0'].size() == torch.Size([16, num_classes])
+
     fake_out = alg.forward_train(fake_input, fake_labels)
+    alg.set_reweight(fake_labels)
     assert fake_out['loss'].item() > 0