+++ tests/test_models/test_algorithms/test_moco.py
@@ -1,10 +1,14 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import platform
+from unittest.mock import MagicMock
+
 import pytest
 import torch
 
+import mmselfsup
 from mmselfsup.models.algorithms import MoCo
 
-queue_len = 8
+queue_len = 32
 feat_dim = 4
 momentum = 0.999
 backbone = dict(
@@ -22,6 +26,19 @@ neck = dict(
 head = dict(type='ContrastiveHead', temperature=0.2)
 
 
+def mock_batch_shuffle_ddp(img):
+    return img, 0
+
+
+def mock_batch_unshuffle_ddp(img, mock_input):
+    return img
+
+
+def mock_concat_all_gather(img):
+    return img
+
+
+@pytest.mark.skipif(platform.system() == 'Windows', reason='Windows mem limit')
 def test_moco():
     with pytest.raises(AssertionError):
         alg = MoCo(backbone=backbone, neck=None, head=head)
@@ -42,3 +59,13 @@ def test_moco():
     assert fake_backbone_out[0].size() == torch.Size([16, 2048, 7, 7])
     with pytest.raises(AssertionError):
         fake_backbone_out = alg.forward_train(fake_input)
+
+    mmselfsup.models.algorithms.moco.batch_shuffle_ddp = MagicMock(
+        side_effect=mock_batch_shuffle_ddp)
+    mmselfsup.models.algorithms.moco.batch_unshuffle_ddp = MagicMock(
+        side_effect=mock_batch_unshuffle_ddp)
+    mmselfsup.models.algorithms.moco.concat_all_gather = MagicMock(
+        side_effect=mock_concat_all_gather)
+    fake_loss = alg.forward_train([fake_input, fake_input])
+    assert fake_loss['loss'] > 0
+    assert alg.queue_ptr.item() == 16